pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ont_tcr_consensus"]

[project]
name = "ONT-TCRconsensus"
version = "0.1.0"
description = "ONT-TCRconsensus creates and counts high accuracy full-length unique TCR molecule consensus sequences."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" } 
authors = [{ name = "Marius Messemaker" },
           { name = "Bjørn P.Y. Kwee" },
        ]

classifiers = [
    "Development Status :: 1 - Planning",
    "Natural Language :: English",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3"
]

dependencies = [
    "bcftools>=1.21",
    "medaka>=2.0.1", 
    "minimap2>=2.24", 
    "numpy>=1.26.4", 
    "pandas>=2.2.2",  
    "pysam>=0.22.0", 
    "pyspoa>=0.2.1", 
    "python-edlib>=1.3.9", 
    "ray-core>=2.37.0", 
    "ray-default>=2.37.0", 
    "samtools>=1.21", 
    "seqkit>=2.8.2", 
    "vsearch>=2.29.0"
]

[project.urls]
Homepage = "https://github.com/schumacherlab/ONT-TCRconsensus"
Repository = "https://github.com/schumacherlab/ONT-TCRconsensus.git"

[project.scripts]
tcr_consensus = "ont_tcr_consensus.tcr_consensus:main"